backend/src/auth/jwt.py
from datetime import datetime, timedelta

from jose import JWTError, jwt

from ..core.config import settings


class JWTError(Exception):
    """Custom exception for JWT errors"""

    pass


def create_access_token(data: dict, expires_delta: int | None = None) -> str:
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now() + timedelta(minutes=expires_delta)
    else:
        expire = datetime.now() + timedelta(
            minutes=settings.access_token_expire_minutes
        )
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(
        to_encode, settings.secret_key, algorithm=settings.algorithm
    )
    return encoded_jwt


def verify_access_token(token: str) -> dict:
    try:
        payload = jwt.decode(
            token, settings.secret_key, algorithms=[settings.algorithm]
        )
        return payload
    except JWTError:
        raise JWTError("Could not validate credentials")


def extract_user_id(token: str) -> dict:
    credentials_exception = JWTError("Could not validate credentials")
    try:
        payload = verify_access_token(token)
        user_id: str = payload.get("user_id")
        if user_id is None:
            raise credentials_exception
        return {"user_id": user_id}
    except JWTError:
        raise credentials_exception
