backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for the Proactive Onboarding Engine (POE) using FastAPI, LangGraph, and pgvector."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "langgraph>=1.0.5",
    "langchain-openai>=1.1.6",
    "httpx>=0.27.0",
    "psycopg[binary]>=3.2.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "duckduckgo-search>=6.1.0",
    "youtube-transcript-api>=1.2.3",
    "pytube>=15.0.0",
    "langchain-community>=0.4.1",
    "deepagents>=0.3.6",
    "pydantic-settings>=2.12.0",
    "structlog>=25.5.0",
    "google-api-python-client>=2.188.0",
    "ddgs>=9.10.0",
    "alembic>=1.18.0",
    "sqlalchemy>=2.0.45",
    "pwdlib[argon2]>=0.3.0",
]
